Avoid crashing shutting down the server
We terminated the process and waited for the server to finish, then we
tried to kill the process but the object was not around anymore
because it was freed as soon as the process finished.

To avoid crashing, we kill the server only if the QProcess object is
still around and the process is still running (meaning that it refused
to exit when we asked it to terminate).

+++ src/xwayland/xwaylandserver.cpp
@@ -153,9 +153,13 @@ void XWaylandServer::shutdown()
     if (m_process) {
         m_process->terminate();
         m_process->waitForFinished();
-        m_process->kill();
-        delete m_process;
-        m_process = nullptr;
+
+        // Kill the process only if it's still running
+        if (m_process && m_process->state() == QProcess::Running) {
+            m_process->kill();
+            delete m_process;
+            m_process = nullptr;
+        }
     }
 }
